+++ go_router/lib/src/go_router_delegate.dart
@@ -11,6 +11,7 @@ import 'go_router_error_page.dart';
 import 'go_router_material.dart';
 import 'go_router_state.dart';
 import 'logging.dart';
+import 'route_data.dart';
 import 'typedefs.dart';
 
 /// GoRouter implementation of the RouterDelegate base class.
@@ -711,9 +712,17 @@ class GoRouterDelegate extends RouterDelegate<Uri>
         pageKey: match.pageKey, // push() remaps the page key for uniqueness
       );
 
-      yield match.route.pageBuilder != null
-          ? match.route.pageBuilder!(context, state)
-          : _pageBuilder(context, state, match.route.builder);
+      final pageBuilder = match.route.pageBuilder;
+
+      Page? page;
+      if (pageBuilder != null) {
+        page = pageBuilder(context, state);
+        if (page is NoOpPage) {
+          page = null;
+        }
+      }
+
+      yield page ?? _pageBuilder(context, state, match.route.builder);
     }
   }