+++ go_router_builder/_go_router_fake/lib/go_router.dart
@@ -1,0 +1,6 @@
+export 'src/route_data.dart';
+
+abstract class BuildContext {
+  /// Navigate to a location.
+  void go(String location, {Object? extra}) => throw UnimplementedError();
+}