+++ go_router_builder/example/test/ensure_build_test.dart
@@ -1,0 +1,16 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+@Tags(['presubmit-only'])
+import 'package:build_verify/build_verify.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  test(
+    'ensure_build',
+    () =>
+        expectBuildClean(packageRelativeDirectory: 'go_router_builder/example'),
+    timeout: const Timeout.factor(3),
+  );
+}